requirements.txt
connexion==2.2.0
python-rocksdb==0.7.0
eth-typing==1.3.0
ethereum-etl==1.3.1
swagger-ui-bundle==0.0.3